update OrderManager below.
1. as OrderManager::flush() would not be called explictly in sf1r-engine,
to get realtime result from getFreqItemSets() and buildFreqItemsets(),
flush "order_db_" in _writeRecord(), so that the new order be fetched in _getOrder().

2. before calling idmlib::AllMaximalSetsSateLite::FindAllMaximalSets(),
it used ItemIndex::getNumItems() to get items count, while this value
might not be updated until ItemIndex::flush() is called, as idmlib::AllMaximalSetsSateLite
relies on this value to allocate space for all items, it would cause
segment falut if the space allocated is not enough.
So it calls ItemManager::maxItemId() to get items count.