fix a bug in IndexTaskService::prepareDocument_().
when SCD is update type and the doc has not been inserted before,
the original code would return true and set doc id to 0,
which would lead to segment fault in search.

now the code is revised to return false when the doc to update has not
been inserted before.