In previous version, the "BOB" and "BOS" rec_types recommend items based on purchase history, they just filter out the items in browse history and shopping cart.
To recommend items related to browse history and shopping cart,
in this commit, the "BOB" rec_type is based on items of browse history,
the "BOS" rec_type is based on items of shopping cart,
and they would filter out the user's purchased items.

In RecommendController::do_recommend(), parameter "session_id" is added for "BOB" rec_type,
so that the items in visit_item() could be used as browse history.