fix a bug in IndexTaskService::doBuildCollection_().
for SCD file of update type, if DOCID (e.x. "123") has not been inserted before,
it would be inserted instead, so the original code revised "isInsert" to
true if its "oldId" is 0, while this might lead to problem, that is,
all docs after DOCID "123" would be insert type, if their DOCID have been
inserted before, they would be ignored as duplicated doc ids and would not
be updated.

Now the code is revised not to change value of "isInsert".