make recommend workers as CollectionHandler members.
in GetRecommendWorkerController::checkWorker(), to get the instance of
GetRecommendWorker, the previous implementation get the instance from
collectionHandler_->recommendSearchService_.

while if the node is configured as recommend worker without master service,
as the "recommendSearchService_" would be NULL, the above implementation
would lead to segment fault.

now it's revised to get recommend worker instance from CollectionHandler
directly, instead of from "recommendSearchService_".