in distributed recommendation, it calls IndexSearchService::getDocumentsByIds() to get document contents.
in previous implementation for SearchWorker::getDocumentsByIds(),
when it fails to get some docs (they are removed, for example),
it would clear all results contents, this would lead the problem that
no docs would be recommended, although some docs exist.

to enable recommending those existing docs, SearchWorker and DocumentManager
are revised that, whenever it fails to get some docs, it would
continue to get those rest docs.