improve the memory cost by AttrTable::valueIdTable_.
in previous implementation, it uses vector<uint32_t> to store attribute
value ids for each doc, it costs about 30 bytes for each doc.

now it's replaced to PropIdTable, which uses 2 vectors to store
attribute value ids for all docs, now it costs about 10 bytes for each doc.

So for 10M docs in b5mp collection, valueIdTable_ would cost 100M bytes.

BTW, the other members in AttrTable cost 700M bytes in b5mp collection,
but this cost does not depend on doc num, it only depends on the number
of attribute names and values.