remove an exception check in test case LogServerConnectionTest/expectHostResolveError.
when the host is empty, LogServerConnection::asynRequest() was
expected to throw exception "std::runtime_error" for unknown host.

as it calls system function "getaddrinfo()" to lookup DNS, on server
172.16.0.161, it's found that sometimes this function returns success
result according to its DNS environment, which makes
LogServerConnection::asynRequest() fail to throw the exception.

as the behaviour is not determined, the excetion check is removed.