Fix bug on ScdParser::iterator::getOffset()
The getOffset() method wrongly returns the offset relative to the
DOCID value instead of the DOCID tag, that is:
<DOCID>[1]23456  instead of [<]DOCID>123456

This happens for all the documents but the first, where the offset
is initialized to 0. For this reason when iterating on an SCD file,
the second document is always discarded, while dereferencing the
iterator on the last document causes a segmentation fault.

A test case has been added.