fix a bug in multi-thread search.
in SearchManager::doSearch_(), previously, it was like below:
1. call pDocIterator->skipTo(docid_start) to skip to a docid (>= docid_start)

2. call while (pDocIterator->next()) to iterate the docs.

there are two problems in above codes:
a. in the 1st step, the current doc id might be a candidate doc, while it was not processed,
   because the 2nd step calls pDocIterator->next() to move to the next doc.

b. in the while loop of the 2nd step, the current doc id might be larger than 'docid_end',
   as it's out of the docid range for that thread, it should not be processed.

now it's revised like below:
a. get the return value from pDocIterator->skipTo().
b. check the docid range in each iteration.