fix a bug of infinite loop in 7c6c5be3979cfdbb8067742e419c0c018611460f.
in previous code of SearchManager::doSearch_(), in the while loop of doc iteration,
if groupFilter->test(curDocId) is false, it calls "continue" to skip this doc,
as "pDocIterator->next()" wasn't called in this case, it leads to infinite loop.

now it's fixed by moving "pDocIterator->next()" into loop condition, so that
when "continue" is called, "pDocIterator->next()" would still be called.