for safe concurrent access on NumericPropertyTableBase in SearchManager.
in each thread, it cannot create read lock twice in the same scope,
that's because if another thread try to create a write lock before the 2nd read lock,
it would lead to dead lock.

now it's revised to insert each NumericPropertyTableBase to a PropSharedLockSet instance,
this instance takes care of the lock and unlock issues, then for each access on
NumericPropertyTableBase, there is no need to lock again, which is also expected for
performance reason.

in summary, below access on NumericPropertyTableBase are inserted to the PropSharedLockSet instance:
1. SortPropertyCache::getComparator()
2. rangePropertyTable
3. counterTables
4. GroupFilter