fix a multi-thread bug in JobScheduler.
this bug appears under below events in sequence:
1. the JobScheduler thread has just executed "asynchronousTasks_.pop(taskCollectionPair)",
   but "runCollectionName_ = taskCollectionPair.second" has not been executed.

2. for the request of "stop_collection", the thread executed "if (collection != runCollectionName_)",
   and its result is true.

in this case, the 1st thread would continue to execute the task,
and the 2nd thread would destroy the bundle of that collection,
then segmentation fault might happen if the 1st thread accesses the memory which has been destroyed.

now this bug is fixed by making the above instructions atomic, that is,
they are under the scope of concurrent_queue's lock.