remove class SortPropertyCache.
previoulsy, SortPropertyCache is used to get the NumericPropertyTableBase instance from
below sources:
1. DocumentManager::getNumericPropertyTable()
2. CTRManager::loadCtrData()
3. IndexManager::loadPropertyDataForSorting() and convertStringPropertyDataForSorting()

as the 3rd point requires PropSharedLock::ScopedWriteLock, it leads to dead lock problem.
as current requirement does not need the 3rd point, and the 1st and 2nd points have been
encapsulated in class NumericPropertyTableBuilder, so we removed class SortPropertyCache,
and use class NumericPropertyTableBuilder instead.