fix a bug when update type SCD has duplicated DOCIDs.
in "sf1-test" project, use below command to run the smoke test for this bug:
$ rspec smoke/sf1/product/F0703UpdateProductPrice.rb

when there are duplicated DOCIDs in update type SCD,
previously, the BTreeIndexer did not remove the docid for old values.

now it's fixed by keeping the first instance of oldIndexDocument,
so that BTreeIndexer could remove the old values properly.