Add query normalization in mapping to top label and custom rank.
Summary:
Previously, the query (mapping to top label and custom rank) is not normalized.
For example, when "iPhone 4S 手机" is mapped to a top label, it would fail to
find label for the query such as "手机 iphone 4s". In order to fix this problem,
the queries (to get/set top label and custom rank) are normalized.

Smoke Test:
sf1-test/smoke/sf1/product/F0702ProductRanking.rb