Fix a bug of no search result in suffix search.
Summary:
In suffix search, sometimes the major tokens got from product matcher is not correct.
For example, when search for query "Haier/海尔 ES50H-D1(E) 50升储水式即热洗澡淋浴电热水器正品",
the major terms was "海尔 d-1 50# 电热水器 es50h-d1", while "d-1" and "50#" didn't appear
in query, it leads to no search result problem in suffix search.

In order to fix this problem, if the token does not appear in query, then it would not be
added as major token.