Improve the performance of AttrCounter.
Replace from std::vector to std::map.

As std::vector's size needs to be the whole number of value ids, which
is 22M for b5mp collection, it costs time to allocate such large
memory (> 100M bytes), so it's changed to std::map, which size would be
the number of value ids from search results.