Improve the performance in StringGroupCounter.
In StringGroupCounter::addDoc(), as the bottleneck lies in calling
PropValueTable::parentIdSet(), it's moved into
StringGroupCounter::getGroupRep() now.

In benchmark test of group by Category on 1M candidate docs,
on local machine, the time cost is changed from 0.21s to 0.088s,
on server machine, the time cost is changed from 0.35s to 0.15s.