Make KNlpDictMonitor support notify with command "rsync".
Previously, it is notified when the timestamp of file "z.update.time" is
changed, while it failed to be notified when using command "rsync"
changing its timestamp.

Now it's revised to monitor the whole directory with the event
"IN_MOVED_TO", the test result shows that it could be notified with
"rsync".

In t_FileMonitor.cpp, a test case "testRSyncDir" is added to test this
feature.