Fix a performance problem in fuzzy search.
In mining, the write lock in CategoryClassifyTable::resize() would block
all the fuzzy search threads which are acquiring the read lock of
CategoryClassifyTable.

From the test, it costs 1 minute in the resize() function to increase
the elements number from 45M to 54M. It means those query threads would
be blocked for 1 minute.

In order to fix this problem, a dedicated "writer instance" is added. That
is, the query thread would access the "reader instance", and the mining
task would access the "writer instance", they would be switched when the
mining task is finished.