Fix a performance problem in Zambezi search.
When the search request has parameters such as "attr_label",
"group_label", or "conditions", it might cost more than 3 seconds in
ZambeziManager::search().

This was caused as izenelib::ir::Zambezi::AttrScoreInvertedIndex would
iterate the postings until it fetches 1M docs which meet the condition
of those labels. In the worst case, it would iterate the whole postings
(more than 10M docs) if the candidate number is less than 1M.

In order to avoid this problem, the filtering is moved after the posting
iteration. That is, it first fetches 1M candidates, then it filters out
those undesirable results.