In zambezi search, move the group count after topk is fetched.
In zambezi iterating postings, it might call ZambeziFilter::test()
multiple times for the same docid, which leads to incorrect group count
values.

In order to avoid this bug, the group count is moved after posting
iteration is finished, so that GroupFilter::test() would be called only
once for each topk docid.