In zambezi search, move the groupFilter after topk docs are fetched.
If the search parameter has "group_label" or "attr_label", in the worst
case, if the candidate number is less than kZambeziTopKNum, it would
iterate the whole postings. This would lead to bad performance.

In order to avoid this problem, the groupFilter is moved after posting
iteration is finished, so that GroupFilter::test() would be called at
most kZambeziTopKNum times.