Improve the performance in condition filter.
Previsouly, FilterCache stores the compressed bitmap, it costs 400ms in
decompressing to 50M bits (in which 25M bits are set).

In order to avoid the decompression time cost on dense bitset,
FilterCache is revised with below policy:

- For dense bitset, store the raw bitset to avoid decompression time.
- For sparse bitset, store the compressed bitset to save memory cost.

Now the dense threshold is choosed as 0.02, that is, if more than 2%
bits are set, it would be assumed as dense.

From the performance test, it shows that the TPS is improved from 37 to
106, and the memory cost is the same (3G bytes).